need to fix minimize with borders

+++ Yamui.Framework/Forms/Glow/YamuiShadowBorder.cs
@@ -3,6 +3,7 @@ using System.Drawing;
 using System.Drawing.Drawing2D;
 using System.Drawing.Imaging;
 using System.Runtime.InteropServices;
+using System.Runtime.Versioning;
 using System.Windows.Forms;
 using Yamui.Framework.Helper;
 
@@ -13,13 +14,11 @@ namespace Yamui.Framework.Forms.Glow {
     /// renders the "glowing effect" on one of the sides
     /// </summary>
     /// <remarks>https://msdn.microsoft.com/en-us/library/windows/desktop/ms633556(v=vs.85).aspx?f=255&MSPPError=-2147217396</remarks>
+    /// <remarks>http://www.nuonsoft.com/blog/2009/05/27/how-to-use-updatelayeredwindow/</remarks>
     /// <remarks>http://simostro.synology.me/simone/2016/04/04/glow-window-effect/</remarks>
-    internal class SideGlow : IDisposable {
+    internal class YamuiShadowBorder : IDisposable, IWin32Window {
         #region private
 
-        private const int ErrorClassAlreadyExists = 1410;
-        private const int AcSrcOver = 0x00;
-        private const int AcSrcAlpha = 0x01;
         private const int Thickness = 9;
 
         private readonly byte[] _alphas = {64, 46, 25, 19, 10, 07, 02, 01, 00};
@@ -35,8 +34,10 @@ namespace Yamui.Framework.Forms.Glow {
 
         private bool _parentWindowIsFocused;
         private Color _activeColor = Color.Cyan;
-        private Color _inactiveColor = Color.LightGray;
+        private Color _inactiveColor = Color.Orange;
         private WinApi.BLENDFUNCTION _blend;
+
+        private WinApi.WNDCLASS _windowClass;
 
         public Point Location { get; private set; }
         public Size Size { get; private set; }
@@ -57,7 +58,7 @@ namespace Yamui.Framework.Forms.Glow {
             }
         }
 
-        internal IntPtr Handle { get; private set; }
+        public IntPtr Handle { get; private set; }
 
         internal bool ParentWindowIsFocused {
             set {
@@ -76,18 +77,21 @@ namespace Yamui.Framework.Forms.Glow {
 
         #region constuctor
 
-        internal SideGlow(DockStyle side, IntPtr parent) {
+        internal YamuiShadowBorder(DockStyle side, IntPtr parent) {
             _side = side;
             _parentHandle = parent;
 
-            _blend = new WinApi.BLENDFUNCTION {
-                BlendOp = AcSrcOver,
-                BlendFlags = 0,
-                SourceConstantAlpha = 255,
-                AlphaFormat = AcSrcAlpha
-            };
+            AppDomain.CurrentDomain.ProcessExit += OnShutdown;
+            AppDomain.CurrentDomain.DomainUnload += OnShutdown;
+
+            _blend = new WinApi.BLENDFUNCTION(255);
             
-            CreateWindow($"GlowSide_{side}_{parent}");
+            CreateWindow();
+        }
+
+        private void OnShutdown(object sender, EventArgs eventArgs) {
+            Close();
+            WinApi.UnregisterClass(_windowClass.lpszClassName, new HandleRef(this, WinApi.GetModuleHandle(null)));
         }
 
         #endregion
@@ -133,51 +137,56 @@ namespace Yamui.Framework.Forms.Glow {
             }
 
             Location = new Point(left, top);
-            WinApi.SetWindowPos(Handle, !IsTopMost ? WinApi.SpecialWindowHandles.HWND_NOTOPMOST : WinApi.SpecialWindowHandles.HWND_TOPMOST, left, top, 0, 0, 0);
-            UpdateZOrder(left, top, WinApi.SetWindowPosFlags.SWP_NOSIZE | WinApi.SetWindowPosFlags.SWP_NOACTIVATE);
+            WinApi.SetWindowPos(Handle, !IsTopMost ? WinApi.SpecialWindowHandles.HWND_NOTOPMOST : WinApi.SpecialWindowHandles.HWND_TOPMOST, left, top, 0, 0, WinApi.SetWindowPosFlags.SWP_NOSIZE | WinApi.SetWindowPosFlags.SWP_NOACTIVATE);
         }
 
-        internal void UpdateZOrder(int left, int top, WinApi.SetWindowPosFlags flags) {
+        private void UpdateZOrder(int left, int top, WinApi.SetWindowPosFlags flags) {
             WinApi.SetWindowPos(Handle, !IsTopMost ? WinApi.SpecialWindowHandles.HWND_NOTOPMOST : WinApi.SpecialWindowHandles.HWND_TOPMOST, left, top, 0, 0, flags);
             WinApi.SetWindowPos(Handle, _parentHandle, 0, 0, 0, 0, NoSizeNoMove | WinApi.SetWindowPosFlags.SWP_NOACTIVATE);
         }
 
         internal void UpdateZOrder() {
-            WinApi.SetWindowPos(Handle, !IsTopMost ? WinApi.SpecialWindowHandles.HWND_NOTOPMOST : WinApi.SpecialWindowHandles.HWND_TOPMOST, 0, 0, 0, 0, NoSizeNoMove | WinApi.SetWindowPosFlags.SWP_NOACTIVATE);
+            //WinApi.SetWindowPos(Handle, !IsTopMost ? WinApi.SpecialWindowHandles.HWND_NOTOPMOST : WinApi.SpecialWindowHandles.HWND_TOPMOST, 0, 0, 0, 0, NoSizeNoMove | WinApi.SetWindowPosFlags.SWP_NOACTIVATE);
             WinApi.SetWindowPos(Handle, _parentHandle, 0, 0, 0, 0, NoSizeNoMove | WinApi.SetWindowPosFlags.SWP_NOACTIVATE);
         }
 
 
         internal void Close() {
             WinApi.CloseWindow(Handle);
-            WinApi.SetParent(Handle, IntPtr.Zero);
-            WinApi.DestroyWindow(new HandleRef(this, Handle));
+            if (Handle != IntPtr.Zero)
+                WinApi.DestroyWindow(new HandleRef(this, Handle));
         }
 
         #endregion
 
         #region private
 
-        private void CreateWindow(string className) {
+        private void RegisterClass() {
             _wndProcDelegate = CustomWndProc;
-
-            // Create WNDCLASS
-            WinApi.WNDCLASS windClass = new WinApi.WNDCLASS {
-                lpszClassName = className,
+            String appDomain = Convert.ToString(AppDomain.CurrentDomain.GetHashCode(), 16);
+            _windowClass = new WinApi.WNDCLASS {
+                lpszClassName = $"{nameof(YamuiShadowBorder)}.{_side}_{_parentHandle}.{VersioningHelper.MakeVersionSafeName(appDomain, ResourceScope.Process, ResourceScope.AppDomain)}",
                 lpfnWndProc = Marshal.GetFunctionPointerForDelegate(_wndProcDelegate)
             };
 
-            var classAtom = WinApi.RegisterClassW(ref windClass);
+            var classAtom = WinApi.RegisterClassW(ref _windowClass);
             int lastError = Marshal.GetLastWin32Error();
-            if (classAtom == 0 && lastError != ErrorClassAlreadyExists) {
-                throw new Exception("Could not register window class");
+            if (classAtom == 0) {
+                throw new Exception("Could not register window class : " + lastError);
             }
-            
+        }
+
+        private void CreateWindow() {
+            _wndProcDelegate = CustomWndProc;
+
+            // Create WNDCLASS
+            RegisterClass();
             var extendedStyle = (
                 WinApi.WindowStylesEx.WS_EX_LEFT |
                 WinApi.WindowStylesEx.WS_EX_LTRREADING |
                 WinApi.WindowStylesEx.WS_EX_RIGHTSCROLLBAR |
                 WinApi.WindowStylesEx.WS_EX_LAYERED |
+                WinApi.WindowStylesEx.WS_EX_TRANSPARENT |
                 WinApi.WindowStylesEx.WS_EX_TOOLWINDOW);
 
             var style = (
@@ -188,8 +197,8 @@ namespace Yamui.Framework.Forms.Glow {
             // Create window
             Handle = WinApi.CreateWindowEx(
                 (int) extendedStyle,
-                className,
-                className,
+                _windowClass.lpszClassName,
+                _windowClass.lpszClassName,
                 (int) style,
                 0,
                 0,
@@ -200,20 +209,14 @@ namespace Yamui.Framework.Forms.Glow {
                 WinApi.NullHandleRef,
                 IntPtr.Zero
             );
-
-            if (Handle == IntPtr.Zero) {
-                return;
-            }
         }
 
         private IntPtr CustomWndProc(IntPtr hWnd, int msg, IntPtr wParam, IntPtr lParam) {
+
             return WinApi.DefWindowProc(hWnd, msg, wParam, lParam);
         }
 
         private void Render() {
-
-
-            // get non client area device
             var screenDc = WinApi.GetDC(WinApi.NullHandleRef);
             if (screenDc == IntPtr.Zero) {
                 return;
@@ -230,7 +233,7 @@ namespace Yamui.Framework.Forms.Glow {
 
                         WinApi.POINT newLocation = new WinApi.POINT(Location);
                         WinApi.SIZE newSize = new WinApi.SIZE(Size);
-                        WinApi.UpdateLayeredWindow(Handle, screenDc, ref newLocation, ref newSize, memDc, ref _ptZero, 0, ref _blend, WinApi.BlendFlags.ULW_ALPHA);
+                        WinApi.UpdateLayeredWindow(Handle, screenDc, ref newLocation, ref newSize, memDc, ref _ptZero, 0, ref _blend, WinApi.BlendingFlags.ULW_ALPHA);
 
                         if (hBitmap != IntPtr.Zero) {
                             WinApi.SelectObject(memDc, hOldBitmap);
@@ -300,17 +303,14 @@ namespace Yamui.Framework.Forms.Glow {
         #region Dispose
 
         public void Dispose() {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        private void Dispose(bool disposing) {
-            if (_disposed) return;
+            if (_disposed) 
+                return;
             _disposed = true;
-            if (Handle == IntPtr.Zero) return;
-
+            if (Handle == IntPtr.Zero) 
+                return;
             WinApi.DestroyWindow(new HandleRef(this, Handle));
             Handle = IntPtr.Zero;
+            GC.SuppressFinalize(this);
         }
 
         #endregion
